Some small refactorings
